pyproject.toml
[project]
name = "rag-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.6",
    "pydantic-ai-slim[outlines-vllm-offline]>=1.37.0",
    "sentence-transformers>=5.2.0",
    "vllm>=0.13.0",
]
